# src/public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"FriendResponseDto": {
				"properties": {
					"friendshipId": {
						"type": "string"
					},
					"user": {
						"properties": {
							"avatarUrl": {
								"type": "string",
								"nullable": true
							},
							"email": {
								"type": "string"
							},
							"lastName": {
								"type": "string"
							},
							"firstName": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"avatarUrl",
							"email",
							"lastName",
							"firstName",
							"id"
						],
						"type": "object"
					},
					"status": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"friendshipId",
					"user",
					"status",
					"createdAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendFriendRequestDto": {
				"properties": {
					"addresseeId": {
						"type": "string"
					}
				},
				"required": [
					"addresseeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"RespondFriendRequestDto": {
				"properties": {
					"action": {
						"type": "string",
						"enum": [
							"accept",
							"decline",
							"block"
						]
					}
				},
				"required": [
					"action"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SharedCartItemDto": {
				"properties": {
					"productName": {
						"type": "string"
					},
					"quantity": {
						"type": "number",
						"format": "double"
					},
					"unitPrice": {
						"type": "number",
						"format": "double"
					},
					"lineTotal": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"productName",
					"quantity",
					"unitPrice",
					"lineTotal"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SharedCartResponseDto": {
				"properties": {
					"id": {
						"type": "string"
					},
					"sharedBy": {
						"properties": {
							"lastName": {
								"type": "string"
							},
							"firstName": {
								"type": "string"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"lastName",
							"firstName",
							"id"
						],
						"type": "object"
					},
					"cart": {
						"properties": {
							"subtotal": {
								"type": "number",
								"format": "double"
							},
							"items": {
								"items": {
									"$ref": "#/components/schemas/SharedCartItemDto"
								},
								"type": "array"
							},
							"id": {
								"type": "string"
							}
						},
						"required": [
							"subtotal",
							"items",
							"id"
						],
						"type": "object"
					},
					"sharedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"sharedBy",
					"cart",
					"sharedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ShareCartDto": {
				"properties": {
					"friendId": {
						"type": "string"
					}
				},
				"required": [
					"friendId"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateUserStatusDto": {
				"properties": {
					"isActive": {
						"type": "boolean"
					},
					"reason": {
						"type": "string",
						"minLength": 5,
						"maxLength": 300
					}
				},
				"required": [
					"isActive"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ModerateProductDto": {
				"properties": {
					"action": {
						"type": "string",
						"enum": [
							"approve",
							"reject",
							"flag"
						]
					},
					"reason": {
						"type": "string",
						"minLength": 5,
						"maxLength": 300
					}
				},
				"required": [
					"action"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "ecommerce-api",
		"version": "1.0.0",
		"description": "E-Commerce REST API — Node.js, Express, TypeScript, TypeORM, PostgreSQL",
		"contact": {}
	},
	"paths": {
...
		"/social/friends/request": {
			"post": {
				"operationId": "SendRequest",
				"responses": {
					"201": {
						"description": "Request sent",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FriendResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Cannot request yourself"
					},
					"409": {
						"description": "Friendship already exists"
					}
				},
				"description": "Send a friend request to another user",
				"tags": [
					"Social"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendFriendRequestDto"
							}
						}
					}
				}
			}
		},
		"/social/friends/{friendshipId}/respond": {
			"put": {
				"operationId": "Respond",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FriendResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Invalid action for current status"
					},
					"403": {
						"description": "Not your request to respond to"
					},
					"404": {
						"description": "Friend request not found"
					}
				},
				"description": "Accept, decline, or block a friend request",
				"tags": [
					"Social"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "friendshipId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/RespondFriendRequestDto"
							}
						}
					}
				}
			}
		},
		"/social/friends": {
			"get": {
				"operationId": "GetFriends",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FriendResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all accepted friends",
				"tags": [
					"Social"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/social/friends/requests": {
			"get": {
				"operationId": "GetPendingRequests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/FriendResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all pending friend requests sent to you",
				"tags": [
					"Social"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/social/cart/share": {
			"post": {
				"operationId": "ShareCart",
				"responses": {
					"201": {
						"description": "Cart shared",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SharedCartResponseDto"
								}
							}
						}
					},
					"400": {
						"description": "Cart is empty"
					},
					"403": {
						"description": "Not friends with this user"
					}
				},
				"description": "Share your cart with an accepted friend",
				"tags": [
					"Social"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ShareCartDto"
							}
						}
					}
				}
			}
		},
		"/social/cart/shared-with-me": {
			"get": {
				"operationId": "GetSharedCarts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/SharedCartResponseDto"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "View all carts that friends have shared with you",
				"tags": [
					"Social"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
...
		"/payments/webhook": {
			"post": {
				"operationId": "HandleWebhook",
				"responses": {
					"200": {
						"description": "OK"
					},
					"400": {
						"description": "Invalid signature"
					}
				},
				"description": "Paystack webhook — handles all payment events.\nNo JWT required — Paystack calls this directly.",
				"tags": [
					"Payments"
				],
				"security": [],
				"parameters": []
			}
		},
...
		"/admin/users": {
			"get": {
				"operationId": "GetAllUsers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"403": {
						"description": "Admins only"
					}
				},
				"description": "Get all users — admin only",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/admin/users/{userId}/status": {
			"post": {
				"operationId": "UpdateUserStatus",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"403": {
						"description": "Admins only"
					},
					"404": {
						"description": "User not found"
					}
				},
				"description": "Suspend or reactivate a user account — admin only",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserStatusDto"
							}
						}
					}
				}
			}
		},
		"/admin/products/{productId}/moderate": {
			"post": {
				"operationId": "ModerateProduct",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"403": {
						"description": "Admins only"
					},
					"404": {
						"description": "Product not found"
					}
				},
				"description": "Approve, reject, or flag a product listing — admin only",
				"tags": [
					"Admin"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "productId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ModerateProductDto"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}